\name{hotROCs-package}
\alias{hotROCs-package}
\alias{hotROCs}
\docType{package}
\title{
\packageTitle{hotROCs}
}
\description{
\packageDescription{hotROCs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hotROCs}
\packageIndices{hotROCs}

To create a matrix of ROC curve areas for variables that might predict
integration site preferences use either \code{\link{ROC.MRC}} for
Matched Random Controls or \code{\link{ROC.ORC}} for Ordinary Random
Controls.  The latter would be used for integration sites isolated by
sonicating or other wise fragmenting the DNA that contains sites. The
former would typically be used for integration sites isolated by using
restriction enzymes.

}

\author{ \packageAuthor{hotROCs}

Maintainer: \packageMaintainer{hotROCs}
}
\keyword{ package }

