% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.DDCP.R
\name{ROC.DDCP}
\alias{ROC.DDCP}
\title{ROC.DDCP}
\usage{
ROC.DDCP(response, indep)
}
\arguments{
\item{response}{logical vector}

\item{indep}{matrix or \code{data.frame} of
\code{length(response)} rows containing numeric values.}
}
\value{
\code{list} with element \code{theta} giving the areas
    under the ROC curve for the variables in \code{indep} and
    element \code{var} giving the variance-covariance matrix of
    the estimates.
}
\description{
ROC curve areas and variances estimated as per Delong et al, 1988.
}
\details{
This is a utility function that users will probably not want to
invoke directly.
}
\seealso{
DeLong, Elizabeth R., David M. DeLong, and Daniel
    L. Clarke-Pearson. \dQuote{Comparing the areas under two or
    more correlated receiver operating characteristic curves: a
    nonparametric approach.}  Biometrics (1988): 837-845.
}
\author{
Charles Berry
}
