% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.MRC.R
\name{ROC.MRC}
\alias{ROC.MRC}
\title{ROC.MRC - ROC areas for matched control data}
\usage{
ROC.MRC(response, stratum, variables, origin = NULL, origin.levels = NULL,
  ragged.OK = TRUE)
}
\arguments{
\item{response}{logical vector identifying cases or character
vector or factor vector with \dQuote{insertion} marking the
cases. Alternatively, a factor or character vector with
elements matching \code{"insertion"} denoting the cases.}

\item{stratum}{a vector of \code{length(response)} elements whose
unique values correspond to cases and their matched
controls. There must be exactly one case and \code{k} controls
in every stratum.}

\item{variables}{numeric a \code{matrix} or \code{data.frame} with
\code{length(response)} rows and two or more columns.}

\item{origin}{\code{NULL} or a vector of \code{length(response)}
elements to identify different data sources.}

\item{origin.levels}{optional character vector of origin levels}

\item{ragged.OK}{logical - \code{TRUE}, if differing numbers of
MRCs of any one origin are acceptable.}
}
\value{
\code{list} with elements \sQuote{ROC} giving a matrix of
    ROC curve areas, \sQuote{var} giving a list of variance
    matrices, and \sQuote{pvalues} giving a list of matrices
    containing pvalues for various contrasts.
}
\description{
ROC curve areas, variances, and p-values for datasets in which
each case has a collection of matching controls.
}
\details{
When data are collected under a scheme in which matched controls
are used to allow for biased selection and every case has \code{k}
controls, the ROC area is estimated as the average fraction of
controls whose variable value is less than that of its reference
case. Variances are computed from the usual sample variance
divided by the number of cases.  It is the responsiibility of the
caller to remove \code{NA} values from the arguments.
}
\examples{
case <- rep(rep(c(TRUE,FALSE),c(1,3)),1000)
group <- rep(1:1000,each=4)
var1 <- case + rnorm(4000)
var2 <- case + rexp(4000)
var3 <- case + runif(4000,-1,1)
origin <- rep(factor(letters[1:4]),each=1000)
ROC.MRC(case,group,cbind(var1,var2,var3),origin)
}
\author{
Charles Berry
}

