% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.ORC.R
\name{ROC.ORC}
\alias{ROC.ORC}
\title{ROC.ORC - ROC area matrix}
\usage{
ROC.ORC(response, variables, origin = NULL, origin.levels = NULL)
}
\arguments{
\item{response}{logical vector identifying cases or character
vector or factor vector with \code{"insertion"} marking the
cases.}

\item{variables}{a \code{matrix} or \code{data.frame} with
\code{length(response)} rows and two or more columns.}

\item{origin}{\code{NULL} or a vector of \code{length(response)}
elements to identify different data sources.}

\item{origin.levels}{optional character vector of origin levels}
}
\value{
\code{list} with elements \sQuote{ROC} giving a matrix of
    ROC curve areas, \sQuote{var} giving a list of variance
    matrices, and \sQuote{pvalues} giving a list of matrices
    containing pvalues for various contrasts.
}
\description{
ROC curve areas, variances, and p-values for datasets in which the
cases have ordinary random controls (i.e. not matched on any
characteristic).  The ROC areas and corresponding variances are
estimated by the method of Delong et al.  The data may derive from
multiple datasets. Comparisons of different variables in a dataset
compare the ROC curve areas relative to the controls.  Comparisons
of different datasets uses only the responses from each dataset
and do not involve the random controls.
}
\details{
A matrix of ROC curve areas and variances.
}
\examples{
case <- rep(rep(c(TRUE,FALSE),c(1,3)),1000)
var1 <- case + rnorm(4000)
var2 <- case + rexp(4000)
var3 <- case + runif(4000,-1,1)
origin <- rep(factor(letters[1:4]),each=1000)
ROC.ORC(case,cbind(var1,var2,var3),origin)
}
\author{
Charles Berry
}

