% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCSVG.R
\name{ROCSVG}
\alias{ROCSVG}
\title{ROC curve heatmaps}
\usage{
ROCSVG(roc.res.list, svg.file.base = "roc")
}
\arguments{
\item{roc.res.list}{An object such as returned by
\code{\link{ROC.ORC}} or \code{\link{ROC.MRC}}}

\item{svg.file.base}{The name of the directory for in which to
store the resulting files}
}
\value{
\code{NULL} -- the function is run for its side effects.
}
\description{
Create an interactive ROC curve heatmap
}
\details{
A file \code{file.path(svg.file.base, "main.svg" ) } is created
 which can be viewed with and SVG viewer like the FireFox
 browser. There are many other files linked to that file that
 allow the user to inspect significance test results.
}
\author{
Charles Berry
}

